Almost working... randomize works.
The lesson learned:
Each Declaration in any Stage's ShaderModule, must have a corresponding
DescriptorSetLayoutBinding (with the binding used in the declaration)
at the index of the .pSetLayouts array of VkPipelineLayoutCreateInfo of
its set_index (as used in the declaration) with the correct DescriptorType
and DescriptorCount with the corresponding bit of that Stage set in
ShaderStageFlags.
Each set.binding slot can only have one DescriptorType/DescriptorCount
pair in the layout (as, likewise, only a single descriptor can be
bound while executing a pipeline).